Remove "as usize" from Reading exact sized arrays

+++ wow_vanilla_messages/src/world/v1/v12/character.rs
@@ -199,7 +199,7 @@ impl Character {
         let pet_familiy = crate::util::read_u32_le(r)?;
 
         // equipment: CharacterGear[19]
-        let mut equipment = Vec::with_capacity(19 as usize);
+        let mut equipment = Vec::with_capacity(19);
         for i in 0..19 {
             equipment.push(CharacterGear::read(r)?);
         }
